pyproject.toml
[tool.poetry]
name = "finanzas-email-tracker"
version = "0.1.0"
description = "Sistema automatizado para rastrear y categorizar transacciones bancarias desde correos de Outlook usando IA"
authors = ["Sebastian Cruz <tu-email@example.com>"]
readme = "README.md"
packages = [{include = "finanzas_tracker", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# ORM y Base de Datos
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
# Validación y Configuración
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0"
python-dotenv = "^1.0.0"
# Microsoft Graph API
msal = "^1.28.0"
msgraph-sdk = "^1.2.0"
# AI y NLP
# Dashboard y UI
streamlit = "^1.38.0"
# Logging
loguru = "^0.7.2"
# Utilidades
python-dateutil = "^2.8.2"
beautifulsoup4 = "^4.12.0"
lxml = "^5.1.0"
anthropic = "^0.72.0"
pypdf = "^4.0.0"
keyring = "^25.7.0"
tenacity = "^8.2.3"
scikit-learn = "^1.7.2"
pdfplumber = "^0.11.8"
psycopg2-binary = "^2.9.11"
asyncpg = "^0.31.0"
fastapi = "^0.122.0"
uvicorn = "^0.38.0"
python-multipart = "^0.0.20"
psycopg = {extras = ["binary"], version = "^3.2.13"}
sentence-transformers = "^5.1.2"
pgvector = "^0.4.1"
# MCP Server
mcp = {version = "^1.22.0", extras = ["cli"]}
pyjwt = "^2.10.1"
bcrypt = ">=4.0,<5.0"
google-api-python-client = "^2.187.0"
google-auth-httplib2 = "^0.2.1"
google-auth-oauthlib = "^1.2.3"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
# Linting y Formatting
ruff = "^0.6.0"
# Type Checking
mypy = "^1.11.0"
# Pre-commit hooks (opcional)
pre-commit = "^3.6.0"
pytest-env = "^1.2.0"
keyrings-alt = "^5.0.2"
testcontainers = "^4.13.3"
types-requests = "^2.32.4.20250913"
types-python-dateutil = "^2.9.0.20251115"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fetch-emails = "finanzas_tracker.scripts.fetch_emails:main"
dashboard = "finanzas_tracker.dashboard.app:main"
mcp-server = "finanzas_tracker.mcp.__main__:main"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

# Solo ignorar dashboard (Streamlit) y mcp (no críticos)
[[tool.mypy.overrides]]
module = [
    "finanzas_tracker.dashboard.*",
    "finanzas_tracker.mcp.*",
]
ignore_errors = true

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["T201", "RUF001"]  # Allow prints and Unicode chars in scripts
"src/finanzas_tracker/services/*.py" = ["RUF001"]  # Allow Unicode chars in services
"src/finanzas_tracker/dashboard/components/*.py" = ["RUF001"]  # Allow Unicode chars in dashboard

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=finanzas_tracker",
    "--cov-report=html",
    "--cov-report=term-missing",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
# Variables de entorno para tests (pytest-env)
env = [
    "USE_POSTGRES=false",
    "DATABASE_URL=sqlite:///:memory:",
    "AZURE_CLIENT_ID=test-client-id",
    "AZURE_TENANT_ID=test-tenant-id",
    "AZURE_CLIENT_SECRET=test-secret",
    "USER_EMAIL=test@example.com",
    "MOM_EMAIL=mom@example.com",
    "ANTHROPIC_API_KEY=sk-ant-test123",
    "ENVIRONMENT=testing",
]

# =============================================================================
# Bandit Security Scanner Config
# =============================================================================
[tool.bandit]
exclude_dirs = ["tests", "alembic", "scripts"]
skips = [
    "B101",  # assert_used - Usamos asserts en tests
    "B104",  # hardcoded_bind_all_interfaces - Necesario para Docker
]

